Fix convolution for splines of even order.
For even-ordered splines, the convolution code multiplies the
surface by -1. The code that does this does not obviously trace
back to any part of the Stroem paper and deleting it gives correct
results for even-ordered splines. It did not run for odd-ordered
splines (as used by IceCube) and so deleting it has no effect there.